refactor: move etcd_metrics_test to transports/etcd.rs and fix integration test
